API 15 update

+++ DragoonMayCry/Record/Model/Extension.cs
@@ -2,10 +2,10 @@ using System.Collections.Generic;
 
 namespace DragoonMayCry.Record.Model
 {
-    public class Extension(string name, ExtensionCategory[] categories, Dictionary<ushort, TrackableDuty> instances)
+    public class Extension(string name, ExtensionCategory[] categories, Dictionary<uint, TrackableDuty> instances)
     {
         public string Name { get; private set; } = name;
         public ExtensionCategory[] Categories { get; private set; } = categories;
-        public Dictionary<ushort, TrackableDuty> Instances { get; private set; } = instances;
+        public Dictionary<uint, TrackableDuty> Instances { get; private set; } = instances;
     }
 }
