+++ DragoonMayCry/State/Tracker/DebuffTracker.cs
@@ -10,7 +10,7 @@ namespace DragoonMayCry.State.Tracker
 {
     internal class DebuffTracker : StateTracker<bool>
     {
-        private readonly Dictionary<ushort, ISet<uint>> debuffInstanceBlacklist = new()
+        private readonly Dictionary<uint, ISet<uint>> debuffInstanceBlacklist = new()
         {
             { 937, new HashSet<uint> { 2935 } }, // Sustained damage on P9S is applied after a TB
             { 939, new HashSet<uint> { 2935 } }, // Failed tower soak on P10S, may not be the player's fault
